Add consistent exceptions
Services and repsoitories now always respond with StatusExceptions if something does not work as expected

+++ src/identity/repositories/UserRepositoryImpl.kt
@@ -1,19 +1,17 @@
 package cozy.identity.repositories
 
+import cozy.exception.middleware.StatusException
 import cozy.identity.data.ClientAuthSigningRequest
+import cozy.identity.extensions.toPemString
+import cozy.identity.services.CertificateService
 import cozy.identity.services.SigningRequestService
 import cozy.services.cluster.data.ClusterUser
-import cozy.identity.services.CertificateService
-import cozy.identity.extensions.toPemString
-import cozy.identity.repositories.SigningRequestRepository
-import cozy.identity.repositories.UserRepository
 import io.fabric8.kubernetes.client.KubernetesClientException
+import io.ktor.http.*
 import kotlinx.coroutines.coroutineScope
 import org.koin.core.component.KoinApiExtension
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
-import java.lang.RuntimeException
-import javax.naming.OperationNotSupportedException
 
 @KoinApiExtension
 class UserRepositoryImpl : UserRepository, KoinComponent {
@@ -25,10 +23,13 @@ class UserRepositoryImpl : UserRepository, KoinComponent {
 
     override suspend fun retrieve(userIdentity: ClusterUser): ClusterUser = coroutineScope {
         val signingRequest = signingRequestRepository.retrieve(userIdentity.id)
-            ?: throw OperationNotSupportedException("Signing Request for '${userIdentity.id}' does not exists.")
+            ?: throw StatusException(
+                HttpStatusCode.NotFound,
+                "Signing Request for '${userIdentity.id}' does not exists."
+            )
 
         if (signingRequest.clientCert == null)
-            throw OperationNotSupportedException("Signing Request has been denied, or is still pending.")
+            throw StatusException(HttpStatusCode.Accepted, "Signing Request has been denied, or is still pending.")
 
         ClusterUser(signingRequest.clientCert!!, signingRequest.clientKey)
     }
@@ -36,7 +37,7 @@ class UserRepositoryImpl : UserRepository, KoinComponent {
     override suspend fun create(userIdentity: ClusterUser): ClusterUser = coroutineScope {
         val keyPair = certificateService.generateKeyPair()
         val clientKey = keyPair.private.toPemString()
-        
+
         val signingRequestData = certificateService.buildSigningRequest(keyPair, userIdentity.toPrincipal())
         val authSigningRequest = ClientAuthSigningRequest(userIdentity.id, clientKey, signingRequestData)
 
@@ -45,18 +46,25 @@ class UserRepositoryImpl : UserRepository, KoinComponent {
             signingRequestService.approve(signingRequest)
 
             retrieve(userIdentity)
-        } catch (e: KubernetesClientException) {
-            throw OperationNotSupportedException("User '${userIdentity.id}' already exists.")
-        } catch (e: RuntimeException) {
+        } catch (e: StatusException) {
+            if (e.cause is KubernetesClientException)
+                throw StatusException(HttpStatusCode.BadRequest, "User '${userIdentity.id}' already exists.")
+
             signingRequestRepository.delete(authSigningRequest.certificateSigningRequest)
 
-            throw OperationNotSupportedException("User '${userIdentity.id}' could not be approved.")
+            throw StatusException(
+                HttpStatusCode.InternalServerError,
+                "User '${userIdentity.id}' could not be approved."
+            )
         }
     }
 
     override suspend fun delete(userIdentity: ClusterUser): Boolean = coroutineScope {
         val signingRequest = signingRequestRepository.retrieve(userIdentity.id)
-            ?: throw OperationNotSupportedException("Signing Request for '${userIdentity.id}' does not exists.")
+            ?: throw StatusException(
+                HttpStatusCode.NotFound,
+                "Signing Request for '${userIdentity.id}' does not exists."
+            )
 
         signingRequestRepository.delete(signingRequest.certificateSigningRequest)
     }
